# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/deposits": {
      "get": {
        "operationId": "DepositsController_list",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Newest first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepositPage"
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ]
      }
    },
    "/api/deposits/{id}": {
      "get": {
        "operationId": "DepositsController_read",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The deposit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deposit"
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ]
      }
    },
...
    "/api/admin/deposit-settings": {
      "get": {
        "operationId": "DepositsAdminController_readSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "In force, or the defaults",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepositSettings"
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ]
      },
      "put": {
        "operationId": "DepositsAdminController_saveSettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositSettingsInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The new version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepositSettings"
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ]
      }
    },
    "/api/admin/deposit-settings/qr": {
      "post": {
        "operationId": "DepositsAdminController_uploadQr",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The re-encoded PNG",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredFileRef"
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ]
      }
    },
    "/api/admin/deposit-settings/qr/{fileId}": {
      "get": {
        "operationId": "DepositsAdminController_qr",
        "parameters": [
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A QR image"
          }
        },
        "tags": [
          "deposits"
        ]
      }
    },
    "/api/admin/deposits": {
      "get": {
        "operationId": "DepositsAdminController_queue",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "default": "submitted",
              "type": "string",
              "enum": [
                "pending",
                "submitted",
                "credited",
                "rejected",
                "expired",
                "cancelled",
                "all"
              ]
            }
          },
          {
            "name": "flagged",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "method",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "sham_cash",
                "usdt_trc20",
                "usdt_bep20"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 254
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rule RV10 order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDepositPage"
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ]
      }
    },
...
    "/api/admin/deposits/{id}": {
      "get": {
        "operationId": "DepositsAdminController_deposit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The deposit and its review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDeposit"
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ]
      }
    },
    "/api/admin/deposits/{id}/receipts/{receiptId}": {
      "get": {
        "operationId": "DepositsAdminController_receipt",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "receiptId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The re-encoded receipt"
          }
        },
        "tags": [
          "deposits"
        ]
      }
    },
    "/api/admin/deposits/{id}/approve": {
      "post": {
        "operationId": "DepositsAdminController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "A UUID per attempt; a retry with the same key returns the first result",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveDeposit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Credited (or its replay)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDeposit"
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ]
      }
    },
    "/api/admin/deposits/{id}/reject": {
      "post": {
        "operationId": "DepositsAdminController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "A UUID per attempt; a retry with the same key returns the first result",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectDeposit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rejected (or its replay)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDeposit"
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ]
      }
    },
    "/api/admin/deposits/{id}/request-receipt": {
      "post": {
        "operationId": "DepositsAdminController_requestReceipt",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestReceipt"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Back to pending",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDeposit"
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ]
      }
    },
    "/api/admin/deposits/{id}/approve-usdt": {
      "post": {
        "operationId": "DepositsAdminController_approveUsdt",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "A UUID per attempt; a retry with the same key returns the first result",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveUsdtDeposit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Credited (or its replay)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDeposit"
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ]
      }
    },
    "/api/admin/deposits/{id}/recheck": {
      "post": {
        "operationId": "DepositsAdminController_recheck",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Verification sent again",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDeposit"
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ]
      }
    },
    "/api/admin/usdt-transfers": {
      "get": {
        "operationId": "DepositsAdminController_transfers",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "method",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "usdt_trc20",
                "usdt_bep20"
              ]
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "default": "unmatched",
              "type": "string",
              "enum": [
                "unmatched",
                "all"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Newest first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUsdtTransferPage"
                }
              }
            }
          }
        },
        "tags": [
          "deposits"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vertex Digital API",
    "description": "",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "DepositMethod": {
        "type": "string",
        "enum": [
          "sham_cash",
          "usdt_trc20",
          "usdt_bep20"
        ]
      },
...
      "Deposit": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "method": {
            "$ref": "#/components/schemas/DepositMethod"
          },
          "status": {
            "$ref": "#/components/schemas/DepositStatus"
          },
          "referenceCode": {
            "type": "string"
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          },
          "declaredAmountUnits": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "declaredUsdUnits": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "quote": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/DepositQuote"
              }
            ]
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "submittedAt": {
            "nullable": true,
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "decidedAt": {
            "nullable": true,
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "rateFixed": {
            "type": "boolean"
          },
          "receiptRequest": {
            "nullable": true,
            "type": "object",
            "properties": {
              "at": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
              },
              "note": {
                "nullable": true,
                "type": "string"
              }
            },
            "required": [
              "at",
              "note"
            ],
            "additionalProperties": false
          },
          "payTo": {
            "nullable": true,
            "type": "object",
            "properties": {
              "accountName": {
                "type": "string"
              },
              "accountNumber": {
                "type": "string"
              },
              "qrUrl": {
                "nullable": true,
                "type": "string"
              }
            },
            "required": [
              "accountName",
              "accountNumber",
              "qrUrl"
            ],
            "additionalProperties": false
          },
          "eta": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ReviewEta"
              }
            ]
          },
          "credited": {
            "nullable": true,
            "type": "object",
            "properties": {
              "usdUnits": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "receivedCurrency": {
                "$ref": "#/components/schemas/Currency"
              },
              "receivedAmountUnits": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "rate": {
                "nullable": true,
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ExchangeRate"
                  }
                ]
              }
            },
            "required": [
              "usdUnits",
              "receivedCurrency",
              "receivedAmountUnits",
              "rate"
            ],
            "additionalProperties": false
          },
          "rejection": {
            "nullable": true,
            "type": "object",
            "properties": {
              "reason": {
                "$ref": "#/components/schemas/DepositRejectReason"
              },
              "note": {
                "nullable": true,
                "type": "string"
              }
            },
            "required": [
              "reason",
              "note"
            ],
            "additionalProperties": false
          },
          "usdt": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/DepositUsdt"
              }
            ]
          }
        },
        "required": [
          "id",
          "method",
          "status",
          "referenceCode",
          "currency",
          "declaredAmountUnits",
          "declaredUsdUnits",
          "quote",
          "expiresAt",
          "createdAt",
          "submittedAt",
          "decidedAt",
          "rateFixed",
          "receiptRequest",
          "payTo",
          "eta",
          "credited",
          "rejection",
          "usdt"
        ],
        "additionalProperties": false
      },
...
      "DepositPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Deposit"
            }
          },
          "nextCursor": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "additionalProperties": false
      },
...
      "DepositSettingsInput": {
        "type": "object",
        "properties": {
          "shamCashAccountName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "shamCashAccountNumber": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "sypEnabled": {
            "type": "boolean"
          },
          "usdEnabled": {
            "type": "boolean"
          },
          "sypQrFileId": {
            "nullable": true,
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "usdQrFileId": {
            "nullable": true,
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "minDepositUsdUnits": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "newAccountPerDepositUsdUnits": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "newAccountDailyUsdUnits": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "establishedPerDepositUsdUnits": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "establishedDailyUsdUnits": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "reviewHoursStart": {
            "$ref": "#/components/schemas/ReviewTime"
          },
          "reviewHoursEnd": {
            "$ref": "#/components/schemas/ReviewTime"
          },
          "reviewTargetMinutes": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1440
          },
          "flagNewAccountUsdUnits": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "flagVelocityCount": {
            "type": "integer",
            "minimum": 1,
            "maximum": 50
          },
          "usdtTrc20Enabled": {
            "type": "boolean"
          },
          "usdtBep20Enabled": {
            "type": "boolean"
          },
          "usdtMinDepositUsdUnits": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "telegramApprovalMaxUsdUnits": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100000000
          }
        },
        "required": [
          "shamCashAccountName",
          "shamCashAccountNumber",
          "sypEnabled",
          "usdEnabled",
          "sypQrFileId",
          "usdQrFileId",
          "minDepositUsdUnits",
          "newAccountPerDepositUsdUnits",
          "newAccountDailyUsdUnits",
          "establishedPerDepositUsdUnits",
          "establishedDailyUsdUnits",
          "reviewHoursStart",
          "reviewHoursEnd",
          "reviewTargetMinutes",
          "flagNewAccountUsdUnits",
          "flagVelocityCount",
          "usdtTrc20Enabled",
          "usdtBep20Enabled",
          "usdtMinDepositUsdUnits",
          "telegramApprovalMaxUsdUnits"
        ]
      },
      "StoredFileRef": {
        "type": "object",
        "properties": {
          "fileId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          }
        },
        "required": [
          "fileId"
        ],
        "additionalProperties": false
      },
...
    }
  }
}